+++ frontend-locadora/src/app/modulos/diretores/containers/diretores/diretores.component.ts
@@ -3,6 +3,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute, Router } from '@angular/router';
 import { catchError, Observable, of } from 'rxjs';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/error-dialog.component';
 
 import { Diretor } from '../../model/diretor';
@@ -52,17 +53,23 @@ export class DiretoresComponent implements OnInit {
   }
 
   onDelete(diretor: Diretor) {
-    this.diretoresService.delete(diretor._id).subscribe({
-      error: () => this.onError('Erro ao tentar remover Diretor.'),
-      complete: () => {
-        this.refresh();
-        this.snackBar.open('Diretor removido com sucesso.', 'X', {
-          duration: 3500,
-          verticalPosition: 'top',
-          horizontalPosition: 'center',
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover este Ator?',
+    });
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.diretoresService.delete(diretor._id).subscribe({
+          error: () => this.onError('Erro ao tentar remover Diretor.'),
+          complete: () => {
+            this.refresh();
+            this.snackBar.open('Diretor removido com sucesso.', 'X', {
+              duration: 3500,
+              verticalPosition: 'top',
+              horizontalPosition: 'center',
+            });
+          },
         });
-      },
+      }
     });
   }
-
 }
